+++ contracts/OrganizationInterface.sol
@@ -12,6 +12,7 @@ interface OrganizationInterface {
     function tally(uint id) external;
     function proposalManager() external view returns (ProposalManagerInterface);
     function votingManager() external view returns (VotingManagerInterface);
+    function quorumReached(uint id) public view returns (bool);
     function winningOption(uint id) public view returns (uint256);
 
 }