+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Bug fix related to proposal times
  - Voting function checking proposal times
  - Passing proposal to VotingRights & Power functions
+ - Update to quorum counting
+ - Replaced ```quorumReached``` with ```maximumQuorum```
  
 ### Removed
  - Vault